Move dataset base class to gammapy.datasets
