+++ gammapy/cube/fit.py
@@ -13,8 +13,9 @@ from gammapy.cube.psf_map import PSFMap
 from gammapy.data import GTI
 from gammapy.irf import EDispKernel, EffectiveAreaTable
 from gammapy.maps import Map, MapAxis
-from gammapy.modeling import Dataset, Parameters
-from gammapy.modeling.models import BackgroundModel, Models, SkyModel
+from gammapy.modeling import Parameters
+from gammapy.datasets import Dataset
+from gammapy.modeling.models import BackgroundModel, Models
 from gammapy.spectrum import SpectrumDataset, SpectrumDatasetOnOff
 from gammapy.stats import cash, cash_sum_cython, wstat
 from gammapy.utils.random import get_random_state
