+++ gammapy/cube/tests/test_fit.py
@@ -15,7 +15,8 @@ from gammapy.cube import (
 from gammapy.data import GTI
 from gammapy.irf import EffectiveAreaTable2D, EnergyDependentMultiGaussPSF
 from gammapy.maps import Map, MapAxis, WcsGeom, WcsNDMap
-from gammapy.modeling import Datasets, Fit
+from gammapy.modeling import Fit
+from gammapy.datasets import Datasets
 from gammapy.modeling.models import (
     BackgroundModel,
     GaussianSpatialModel,
