+++ gammapy/spectrum/dataset.py
@@ -6,7 +6,7 @@ from astropy.io import fits
 from astropy.table import Table
 from gammapy.data import GTI
 from gammapy.irf import EDispKernel, EffectiveAreaTable, IRFStacker
-from gammapy.modeling import Dataset, Parameters
+from gammapy.datasets import Dataset
 from gammapy.modeling.models import Models, SkyModel
 from gammapy.stats import cash, significance, significance_on_off, wstat
 from gammapy.utils.fits import energy_axis_to_ebounds
